requirements.txt
fastapi==0.115.6
langchain==0.3.12
langchain_chroma==0.1.4
langchain_core==0.3.25
langchain_google_genai==2.0.7
langchain_text_splitters==0.3.3
langgraph==0.2.59
pydantic==2.10.3
python-dotenv==1.0.1
typing_extensions==4.12.2
